+++ src/main/kotlin/com/github/sleepypanda/feesh/FeeshMod.kt
@@ -77,7 +77,7 @@ class FeeshMod : ClientModInitializer {
         SeaCreaturesCocoonPublisher.init()
         RareDropsPublisher.init()
         PartyChatPublisher.init()
-        TrophyDiscoveredPublisher.init()
+        TrophyCatchPublisher.init()
         PetLevelUpPublisher.init()
         SacksItemPickupPublisher.init()
         InventoryItemPickupPublisher.init()
@@ -126,6 +126,7 @@ class FeeshMod : ClientModInitializer {
         LootshareMessage.init()
         TrophyFrogDiscoveredMessage.init()
         TrophyFishDiscoveredMessage.init()
+        HideTrophyCatchMessages.init()
 
         // PB
         DoubleHookPersonalBest.init()